pyproject.toml
[project]
name = "cardcounter"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.5,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "openai (>=1.78.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
